requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
auto-py-to-exe==2.46.0
av==14.4.0
bleak==1.0.1
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comfyui-embedded-docs==0.2.6
comfyui_frontend_package==1.25.11
comfyui_workflow_templates==0.1.75
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
diffusers
distro==1.9.0
dlib
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
Eel==0.18.2
einops==0.8.1
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gast==0.4.0
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-crc32c==1.7.1
google-pasta==0.2.0
gradio==5.42.0
gradio_client==1.11.1
greenlet==3.2.4
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
hugchat==0.5.1
huggingface-hub==0.34.4
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
invisible-watermark==0.2.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jax==0.7.0
jaxlib==0.7.0
jcloud==0.3
jedi==0.19.2
jina-hubble-sdk==0.39.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
keras==3.11.2
keyboard==0.13.5
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lap==0.5.12
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
namex==0.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
openai==1.99.3
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pefile==2023.2.7
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pylibsrtp==0.12.0
PyMsgBox==1.0.9
pymunk==7.1.0
pyngrok==7.3.0
pynput==1.8.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.9.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.1
qrcode==8.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
spandrel==0.4.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchsde==0.2.6
torchvision==0.22.1+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.174
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
wrapt==1.17.2
xformers==0.0.31.post1
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2